+++ src/components/package-search-list.js
@@ -35,6 +35,8 @@ function PackageSearchList({
           active={item.content && activeId && item.content.id === activeId}
           shouldFocus={item.content && shouldFocusItem && item.content.id === shouldFocusItem}
           onClick={() => onClickItem(`/eholdings/packages/${item.content.id}`)}
+          showProviderName
+          showTitleCount
         />
       )}
     />