+++ src/components/settings/settings-access-status-types/settings-access-status-types.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Field } from 'redux-form';
 import {
@@ -21,6 +21,8 @@ import { EditableList } from '@folio/stripes/smart-components';
 
 import Toaster from '../../toaster';
 
+import { accessStatusTypeDataShape } from '../../../constants/accessTypesReduxStateShape';
+
 const SettingsAccessStatusTypes = ({
   accessTypesData,
   onCreate,
@@ -39,6 +41,29 @@ const SettingsAccessStatusTypes = ({
     attributes,
   });
 
+  useEffect(() => {
+    const errorsLength = accessTypesData.errors.length;
+
+    if (errorsLength) {
+      const lastErrorTitle = accessTypesData.errors[errorsLength - 1].title;
+
+      if (lastErrorTitle.endsWith('not found')) {
+        setToasts(currentToasts => [
+          ...currentToasts,
+          {
+            id: `access-type-delete-failure-${Date.now()}`,
+            message: <FormattedMessage id="ui-eholdings.settings.accessStatusTypes.delete.error" />,
+            type: 'error',
+          }
+        ]);
+
+        setSelectedStatusType(null);
+        setShowConfirmDialog(false);
+      }
+    }
+  }, [accessTypesData.errors]);
+
+
   const formatter = {
     name: ({ attributes }) => (attributes?.name ?? <NoValue />),
     description: ({ attributes }) => (attributes?.description ?? <NoValue />),
@@ -56,7 +81,7 @@ const SettingsAccessStatusTypes = ({
       ) : <NoValue />;
     },
     // records will be done after MODKBEKBJ-378
-    records: () => <NoValue />
+    records: ({ usageNumber }) => usageNumber || <NoValue />
   };
 
   // eslint-disable-next-line react/prop-types
@@ -111,6 +136,10 @@ const SettingsAccessStatusTypes = ({
         id="ui-eholdings.settings.accessStatusTypes.validation"
         values={{ limit: 75 }}
       />;
+    }
+
+    if (value && accessTypesData.items.find(accessType => accessType.attributes.name === value)) {
+      return <FormattedMessage id="ui-eholdings.settings.accessStatusTypes.validation.duplicate" />;
     }
 
     return null;
@@ -194,6 +223,7 @@ const SettingsAccessStatusTypes = ({
             onUpdate={onUpdate}
             readOnlyFields={['lastUpdated', 'records']}
             visibleFields={['name', 'description', 'lastUpdated', 'records']}
+            actionSuppression={{ delete: accessType => accessType.usageNumber, edit: () => false }}
           />
         )}
       </IntlConsumer>
@@ -222,25 +252,10 @@ const SettingsAccessStatusTypes = ({
 
 SettingsAccessStatusTypes.propTypes = {
   accessTypesData: PropTypes.shape({
+    errors: PropTypes.arrayOf(PropTypes.objectOf(PropTypes.string)).isRequired,
     isDeleted: PropTypes.bool.isRequired,
     isLoading: PropTypes.bool.isRequired,
-    items: PropTypes.arrayOf(PropTypes.shape({
-      attributes: PropTypes.shape({
-        description: PropTypes.string,
-        name: PropTypes.string.isRequired,
-      }).isRequired,
-      creator: PropTypes.objectOf(PropTypes.string),
-      id: PropTypes.string.isRequired,
-      metadata: PropTypes.shape({
-        createdByUserId: PropTypes.string.isRequired,
-        createdByUsername: PropTypes.string.isRequired,
-        createdDate: PropTypes.string.isRequired,
-        updatedByUserId: PropTypes.string,
-        updatedDate: PropTypes.string,
-      }),
-      type: PropTypes.string.isRequired,
-      updater: PropTypes.objectOf(PropTypes.string),
-    })),
+    items: PropTypes.arrayOf(accessStatusTypeDataShape),
   }),
   confirmDelete: PropTypes.func.isRequired,
   onCreate: PropTypes.func.isRequired,