+++ src/components/title-search-filters.js
@@ -1,7 +1,11 @@
 import React from 'react';
-import { FormattedMessage } from 'react-intl';
 
 import SearchFilters from './search-form/search-filters';
+import {
+  titleSortFilterConfig,
+  selectionStatusFilterConfig,
+  publicationTypeFilterConfig,
+} from '../constants';
 
 /**
  * Renders search filters with specific title filters.
@@ -15,46 +19,11 @@ function TitleSearchFilters(props) {
   return (
     <SearchFilters
       searchType="titles"
-      availableFilters={[{
-        name: 'sort',
-        label: <FormattedMessage id="ui-eholdings.label.sortOptions" />,
-        defaultValue: 'relevance',
-        options: [
-          { label: <FormattedMessage id="ui-eholdings.filter.sortOptions.relevance" />, value: 'relevance' },
-          { label: <FormattedMessage id="ui-eholdings.label.title" />, value: 'name' }
-        ]
-      }, {
-        name: 'selected',
-        label: <FormattedMessage id="ui-eholdings.label.selectionStatus" />,
-        defaultValue: 'all',
-        options: [
-          { label: <FormattedMessage id="ui-eholdings.filter.all" />, value: 'all' },
-          { label: <FormattedMessage id="ui-eholdings.selected" />, value: 'true' },
-          { label: <FormattedMessage id="ui-eholdings.notSelected" />, value: 'false' },
-          { label: <FormattedMessage id="ui-eholdings.filter.selectionStatus.orderedThroughEbsco" />, value: 'ebsco' }
-        ]
-      }, {
-        name: 'type',
-        label: <FormattedMessage id="ui-eholdings.label.publicationType" />,
-        defaultValue: 'all',
-        options: [
-          { label: <FormattedMessage id="ui-eholdings.filter.all" />, value: 'all' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.audioBook" />, value: 'audiobook' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.book" />, value: 'book' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.bookSeries" />, value: 'bookseries' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.database" />, value: 'database' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.journal" />, value: 'journal' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.newsletter" />, value: 'newsletter' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.newspaper" />, value: 'newspaper' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.proceedings" />, value: 'proceedings' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.report" />, value: 'report' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.streamingAudio" />, value: 'streamingaudio' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.streamingVideo" />, value: 'streamingvideo' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.thesisdissertation" />, value: 'thesisdissertation' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.website" />, value: 'website' },
-          { label: <FormattedMessage id="ui-eholdings.filter.pubType.unspecified" />, value: 'unspecified' }
-        ]
-      }]}
+      availableFilters={[
+        titleSortFilterConfig,
+        selectionStatusFilterConfig,
+        publicationTypeFilterConfig,
+      ]}
       {...props}
     />
   );