+++ src/constants/index.js
@@ -14,3 +14,4 @@ export { default as accessTypesReduxStateShape } from './accessTypesReduxStateSh
 export { default as KbCredentials } from './kbCredentialsReduxStateShape';
 export { default as KbCredentialsUsers } from './kbCredentialsUsersReduxStateShape';
 export { default as rootProxy } from './rootProxyReduxStateShape';
+export * from './filterConfigs';