+++ test/bigtest/tests/provider-show-test.js
@@ -62,7 +62,7 @@ describe('ProviderShow', () => {
     });
 
     it('displays number of selected titles for a package', () => {
-      expect(ProviderShowPage.packageList(0).numTitlesSelected).to.equal(`${packages[0].selectedCount}`);
+      expect(ProviderShowPage.packageList(0).numTitlesSelected).to.equal(`Selected: ${packages[0].selectedCount}`);
     });
 
     it('displays total number of titles for a package', () => {