+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Change history for ui-eholdings
 
-## [3.1.0] (IN PROGRESS)
+## [4.0.0] (IN PROGRESS)
+
+* Pin `moment` at `~2.24.0`. Refs STRIPES-678.
+* Add filtering of titles by Access Types. (UIEH-834)
+* Add validation for duplicated Access Types. (UIEH-878)
+* Add permission names to translations. (UIEH-890)
+* Update react-intl to v4
+* Update `stripes` to `v4`
+* Hide delete button when access type is used in records. (UIEH-868)
+* Provider Record list of Packages UI updates. (UIEH-861)
+* Title Record list of Packages UI updates. (UIEH-863)
+* Ability to create multiple Knowledge base configurations. (UIEH-856)
+* Add a Name field to Settings Knowledge Base page. (UIEH-855)
+
+## [3.0.2] (https://github.com/folio-org/ui-eholdings/tree/v3.0.2) (2020-04-08)
 
 * Retain search within packages search and filter selections on the Provider record (UIEH-847)
 * Retain search within titles search and filter selections on the Package record (UIEH-853)
 * Display Access Status Types on Create/Edit/View records. (UIEH-831)
-* Ability to create multiple Knowledge base configurations. (UIEH-856)
-* Add a Name field to Settings Knowledge Base page. (UIEH-855)
 
 ## [3.0.1] (https://github.com/folio-org/ui-eholdings/tree/v3.0.1) (2020-04-02)